switch frequency table to double

+++ src/hertz.c
@@ -36,7 +36,7 @@ foreach my $midi_note (0 .. 127) {
 
  **/
 
-const float hertz[] = {
+const double hertz[] = {
 	    8.1757989156, //   0
 	    8.6619572180, //   1
 	    9.1770239974, //   2
