+++ src/hertz.h
@@ -23,6 +23,6 @@
 #ifndef _HERTZ_H_
 #define _HERTZ_H_
 
-extern const float hertz[];
+extern const double hertz[];
 
 #endif
